requirements.txt
fastapi==0.111.0
email-validator>=2.0
uvicorn[standard]==0.30.1
mangum==0.17.0
httpx>=0.24
pydantic>=1.10,<3
python-dotenv>=0.21
hydra-db-python==0.1.5
